# user-api/swagger.yaml
openapi: 3.0.0
info:
  title: User Management API
  description: API for managing users in a MongoDB database with CRUD operations.
  version: '1.0.0'
servers:
  - url: http://localhost:3000/
paths:
  /users:
    get:
      summary: Retrieves a list of users
      responses:
        '200':
          description: A JSON array of user objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Creates a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
  /users/{userId}:
    get:
      summary: Retrieves a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Updates a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
    delete:
      summary: Deletes a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
components:
  schemas:
    User:
      type: object
      required:
        - firstName
        - lastName
        - placeOfBirth
        - dateOfBirth
        - gender
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        placeOfBirth:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum:
            - Male
            - Female
            - Other